+++ e2e-tests/mdx/src/pages/fs-api/plus+and+static+query/{Mdx.fields__slug}.js
@@ -1,0 +1,25 @@
+import React from "react"
+import { graphql, useStaticQuery } from "gatsby"
+
+export default function FSAPIWithPlusComponent(props) {
+  const data = useStaticQuery(graphql`
+    {
+      site {
+        siteMetadata {
+          plus: title
+        }
+      }
+    }
+  `)
+  return (
+    <>
+      <pre>
+        <code data-cy="static-query-result">
+          {data?.site?.siteMetadata?.plus}
+        </code>
+      </pre>
+      <hr />
+      {props.children}
+    </>
+  )
+}