+++ e2e-tests/production-runtime/src/components/footer.js
@@ -1,9 +1,24 @@
 import React, { useContext } from "react"
+import { useStaticQuery, graphql } from "gatsby"
 import { AppContext } from "../app-context"
 
 // Use as a Slice
-function Footer({ framework, lang, sliceContext: { framework: frameworkViaContext }}) {
+function Footer({
+  framework,
+  lang,
+  sliceContext: { framework: frameworkViaContext },
+}) {
   const { posts } = useContext(AppContext)
+
+  const data = useStaticQuery(graphql`
+    {
+      site {
+        siteMetadata {
+          alias: title
+        }
+      }
+    }
+  `)
 
   return (
     <footer
@@ -12,10 +27,16 @@ function Footer({ framework, lang, sliceContext: { framework: frameworkViaContex
         fontSize: `12px`,
       }}
     >
-     <span data-testid="footer-slice-context-value">{frameworkViaContext}</span>
-     <span data-testid="footer-static-text">Built with {` `}</span>
-     <span data-testid="footer-props">{`${framework}${lang}`}</span>
-     {` `}Posts Count: <span data-testid="footer-context-derieved-value">{`${posts.length}`}</span>
+      <span data-testid="footer-slice-context-value">
+        {frameworkViaContext}
+      </span>
+      <span data-testid="footer-static-text">Built with {` `}</span>
+      <span data-testid="footer-props">{`${framework}${lang}`}</span>
+      {` `}Posts Count:{" "}
+      <span data-testid="footer-context-derieved-value">{`${posts.length}`}</span>
+      <span data-testid="footer-static-query-title">
+        {data.site.siteMetadata.alias}
+      </span>
     </footer>
   )
 }