+++ integration-tests/artifacts/src/pages/react-lazy.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+const LazyComponent = React.lazy(() =>
+  import("../components/react-lazily-imported.js")
+)
+
+export default function PageUsingReactLazyComponent() {
+  return <LazyComponent />
+}