+++ integration-tests/ssr/gatsby-node.js
@@ -1,0 +1,6 @@
+exports.createPages = ({ actions }) => {
+  actions.createSlice({
+    id: `test`,
+    component: require.resolve(`./src/components/test-slice`),
+  })
+}