+++ packages/gatsby-plugin-mdx/src/__tests__/remark-infer-toc-meta.ts
@@ -10,7 +10,13 @@ Some text with *formatting*.
 
 ## Headline 2
 
-With some text beneath`
+With some text beneath
+
+## This heading has **bold** and *italicized* text
+
+## This heading has \`inline code\`
+
+## This heading contains a [link](#)`
 
 describe(`remark: infer ToC meta`, () => {
   it(`parses ToC and attaches it to our meta object`, async () => {
@@ -29,6 +35,18 @@ describe(`remark: infer ToC meta`, () => {
               {
                 title: `Headline 2`,
                 url: `#headline-2`,
+              },
+              {
+                title: `This heading has bold and italicized text`,
+                url: `#this-heading-has-bold-and-italicized-text`,
+              },
+              {
+                title: `This heading has inline code`,
+                url: `#this-heading-has-inline-code`,
+              },
+              {
+                title: `This heading contains a link`,
+                url: `#this-heading-contains-a-link`,
               },
             ],
             title: `Headline`,