+++ packages/gatsby-source-shopify/types/interface.d.ts
@@ -147,7 +147,8 @@ interface IShopifyPluginOptions {
   shopifyConnections: Array<string>
   typePrefix: string
   salesChannel: string
-  prioritize?: boolean
+  prioritize?: boolean,
+  apiVersion: string
 }
 
 interface IGraphQLClient {